Renamed careuser components and updated imports
