+++ Singer.API/ClientApp/src/app/modules/admin/admin.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { AdminRoutingModule } from './admin-routing.module';
-import { OverviewComponent } from './components/careusers/overview/overview.component';
+import { CareUserOverviewComponent } from './components/careusers/careuser-overview/care-user-overview.component';
 import { MaterialModule } from '../../material.module';
 import { CareUserProxy } from '../core/services/care-users-api/careuser.proxy';
 import { CareUserService } from '../core/services/care-users-api/careusers.service';
 import { ApiService } from '../core/services/api.service';
 import { AgegroupPipe } from '../core/services/agegroup.pipe';
 import { CoreModule } from '../core/core.module';
-import { CareUserDetailsComponent } from './components/careusers/care-user-details/care-user-details.component';
+import { CareUserDetailsComponent } from './components/careusers/careuser-details/care-user-details.component';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { KeysPipe } from '../core/services/keys.pipe';
 import {
@@ -45,7 +45,7 @@ import { PendingRegistrationsComponent } from './components/pending-registration
 
 @NgModule({
    declarations: [
-      OverviewComponent,
+      CareUserOverviewComponent,
       CareUserDetailsComponent,
       LegalguardianOverviewComponent,
       LegalguardianDetailsComponent,
