+++ Singer.API/ClientApp/src/app/modules/admin/components/careusers/careuser-details/care-user-details.component.spec.ts
